fix(converter): the CLI never ran, and the privacy guarantee had a hole the array check could not see
Adversarial review of this PR. Two findings that matter, both of them the kind
this repo has a written rule about.

**The documented command did nothing.** The logic and the CLI shared one module,
with the entry point guarded by
`if (process.argv[1]?.includes('trainer_take_to_fixture'))` so a test importing
the module would not run it. Under `vite-node` — the documented way to run it —
`argv[1]` is `node_modules/.bin/vite-node` and the script's own path never
appears in `argv` at all. So the guard was always false: the command printed
nothing and exited 0. Every unit test passed throughout, because they import
`convertTake` directly.

That is exactly the shape CLAUDE.md's "a feature nobody can find is not shipped"
rule names, committed in the same PR whose body cites that rule. It was caught by
a reviewer actually running the documented command, which is the only check that
would have.

Fixed by the split this repo already uses — `lib_audio.ts` holds logic and is
imported by tests, `render_audio.ts` is the CLI and calls `main()`
unconditionally. So: `scripts/lib_trainer_take.ts` (pure) and
`scripts/trainer_take_to_fixture.ts` (CLI). Nothing has to detect whether it is
the entry module, because only the CLI file has a side effect.
`test/trainer_take_cli.test.ts` runs the documented invocation and asserts it
produces output — reachability, not logic.

**The array check could not see flattened coordinates.** `carriesLandmarks` finds
`landmarks`/`keypoints` ARRAYS, but the pinned vectors are a flat
`Record<string, number>` — and `face.head.x`/`face.head.y` are the nose tip's
normalized image coordinates (`src/features/face_catalog.ts:256`, computed as
`lx(ctx.landmarks, FL.noseTip)`). Over a take that is a frame-by-frame nose
trajectory, and the generated README asserted "No landmark or keypoint geometry"
over the top of it.

The good outcome is that stripping them costs nothing: `FACE_OMIT` is the
trainer's OWN declaration that the learner does not use those four features. They
were recorded only because `routineGroups` unions a routine's demanded groups
without applying each cue's `omit` (`src/enroll/cue.ts`), so the engine computes a
superset of what the learner consumes. The fixture should carry what was used.

So they are stripped by feature id, sourced from `FACE_OMIT` and the catalog's
`hand.position.raw` group rather than a hand-written list — and the removal is
REPORTED in the result, `meta.json` and the README, because "a converter that
silently drops a field is one nobody checks" has to apply to this converter too.

Also broadens the stale-stream fix: clearing only the gzip counterpart missed the
general case, where a re-record producing FEWER keys (no hands in frame) left the
old key's file behind and `loadStream` kept serving the previous take's data while
`meta.json` correctly omitted it. The whole directory's streams are now cleared
before writing.

And two doc corrections: `meta.json` is emitted by three scripts now, not two, and
TESTING.md's stale suite counts.

Every new guard mutation-tested. 1377 -> 1384 tests.
