+++ CLAUDE.md
@@ -246,7 +246,7 @@ So, when you add a user-facing capability:
 | **Feature demand** (#163) — a non-Lab consumer claims feature GROUPS; the vector nodes (and the face-model gate) compute them with the Lab closed | `src/features/demand.ts`, `src/app/featureDemand.ts` |
 | Legacy app (**frozen**) | `src/App.tsx`, `src/components/`, `src/hooks/`, `src/plugins/ai-dj/` |
 | Fixtures + replay | `test/fixtures/`, `scripts/record_stream.ts`, `src/dag/recorder.ts`, `test/helpers/fixtures.ts` (the one loader) |
-| **Trainer take → fixture** — the cue interval is the ground truth a clip cannot supply; refuses landmark geometry | `scripts/trainer_take_to_fixture.ts`, `docs/TESTING.md` |
+| **Trainer take → fixture** — the cue interval is the ground truth a clip cannot supply; refuses landmark geometry | `scripts/lib_trainer_take.ts` (logic) + `scripts/trainer_take_to_fixture.ts` (CLI), `docs/TESTING.md` |
 | Conceptual model | `docs/design/component-model.md` |
 
 ## Roadmap & tracking
